+++ Source/SwiftLintBuiltInRules/Rules/Idiomatic/PreferKeyPathRule.swift
@@ -29,6 +29,7 @@ struct PreferKeyPathRule: OptInRule {
             Example("[1, 2, 3].reduce(1) { $0 + $1 }", configuration: extendedMode),
             Example("f.map(1) { $0.a }"),
             Example("f.filter({ $0.a }, x)"),
+            Example("#Predicate { $0.a }"),
         ],
         triggeringExamples: [
             Example("f.map ↓{ $0.a }"),
@@ -174,10 +175,9 @@ private extension ClosureExprSyntax {
     }
 
     func isInvalid(restrictToStandardFunctions: Bool) -> Bool {
-        if keyPathInParent == \FunctionCallExprSyntax.calledExpression {
-            return true
-        }
-        if parent?.is(MultipleTrailingClosureElementSyntax.self) == true {
+        guard keyPathInParent != \FunctionCallExprSyntax.calledExpression,
+              let parentKind = parent?.kind,
+              ![.macroExpansionExpr, .multipleTrailingClosureElement].contains(parentKind) else {
             return true
         }
         if let call = parent?.as(LabeledExprSyntax.self)?.parent?.parent?.as(FunctionCallExprSyntax.self) {